Clean IO and fixture code
